making js script for CRUD operation for inventaris and infaw jumat feature

+++ database/migrations/2025_10_29_030306_create_infaq_jumat_table.php
@@ -14,7 +14,8 @@ return new class extends Migration
         Schema::create('infaq_jumat', function (Blueprint $table) {
             $table->uuid('id_infaq_jumat')->primary();
             $table->date('tanggal');
-            $table->int('nominal');
+            // PERBAIKAN: Mengganti $table->int('nominal') menjadi $table->bigInteger('nominal')
+            $table->bigInteger('nominal'); 
             $table->timestamps();
         });
     }
